services/orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator"
version = "0.1.0"
description = "Music transcription orchestrator service"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "asyncpg>=0.29.0",
]
requires-python = ">=3.9"
